Phoneme Features and Licence
Changed phone_features.py to phoneme_features.py and made it a class
which is held in the utt or dictionary when passed around instead of
callable methods.

Should allow for easy integration of CMUdict when support is added.
(This probably contains errors as I have not tested all the changes very
well)